requirements.txt
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
djangorestframework==3.9.1
google-api-core==1.7.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-translate==1.3.3
googleapis-common-protos==1.5.8
idna==2.8
protobuf==3.6.1
psycopg2==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
